+++ src/libcoev/event.h
@@ -17,8 +17,8 @@ namespace coev
 	{
 		chain *m_eventchain = nullptr;
 		std::coroutine_handle<> m_awaiting = nullptr;
-		uint32_t m_tag = 0;
-		event(chain *obj, uint32_t _tag);
+		uint64_t m_tag = 0;
+		event(chain *obj, uint64_t _tag);
 		event(chain *obj);
 		virtual ~event();
 		event(event &&) = delete;
