+++ src/libcoev/ossignal.h
@@ -15,14 +15,14 @@ namespace coev
 {
 	struct ossignal final : EVRecv
 	{	
-		ossignal(uint32_t id);
+		ossignal(uint64_t id);
 		virtual ~ossignal();
 		int resume();
 		int max() const { return SIGNALMAX; }
 		int min() const { return SIGNALMIN; }
 
-		uint32_t m_id = 0;
-		uint32_t m_tag = 0;
+		uint64_t m_id = 0;
+		uint64_t m_tag = 0;
 		struct ev_signal m_Signal;
 		static void cb_signal(struct ev_loop *loop, struct ev_signal *w, int revents);
 	};
